Made the changes to get a Triangle drawing with Metal
Added a Metal Shader with some simple functions.
Adjusted the Scene to work with the Metal Context.
Moved the Inflight Semaphore and the Command Buffer out of MTLWindow to
MTLGraphicSystem.
Changed to passing an NSWindow instead of an NSView to MTLWindow making
the Metal Layer setup process much simpler.
Made other changes to get basic functionality working.