Added Ubiquity Container to IOSFileSystem to enable iCloud support.
IOSFileSystem now handles setup and initialization of a Ubiquity
Container for managing files on Apple’s iCloud platform.
Added a Condition class to act as an object oriented Condition Variable.
Split the indexing functionality in Semaphore to it’s own class
IndexedSemaphore.
Refactored the other files using Indexing with Semaphore.